Fix Concourse CI lint job
The golangci/golangci-lint image was recently updated from v1.52 to
v1.53 which caused our Concourse CI lint job to start failing. It
seems like this new image does not play well with our prod Concourse
workers since the new image seems to work on our developer Concourse
instances that are on newer versions of Concourse. To work around the
issue for now, pin the image to v1.52.